Allow retrieval of the iterator_range of a position_tagged.
position_cache allows to retrieve the iterator_range of a
position tagged AST type.  However, pos_cache in error_handler<Iterator>
is private, so the iterator_range can not be retrieved from client code.
It would be useful for a "late raw" and for testing purposes.
This patch adds error_handler<Iterator>::position_of(position_tagged).