attribute type specification used `basic_string`
It was [brought to my attention](http://stackoverflow.com/questions/11421430/string-parser-with-boost-variant-recursive-wrapper/11422737?noredirect=1#comment49970878_11422737) on SO that the documentation seems to be (confusingly) in error.

I admit I get more questions about "why can't I assign the synthesized attribute to a `std::string`" than I could explain. So I guess this should be fixed.