[MSVC] problem with BOOST_FUSION_DEFINE_STRUCT_INLINE
There is a bug in MSVC that it will call non-explicit constructor of a destination type instead of calling type conversion operator of source class.

BOOST_FUSION_DEFINE_STRUCT_INLINE generates a class that has non-explicit template constructor calling boost::fusion::copy. MSVC instantianes it and fails in boost::fusion::copy because boost::value_initialized<T> is not a sequence.