X3: improve error handler bug-fix
error_handler::get_line_start now returns an iterator to the actual start
of the line where an error occurred instead of an iterator to the line
break of the previous line (previous line end). The only case where
method returned an iterator to the line start was when the error
occurred in the first line of the input.

Call sites have been adjusted to accommodate the changes.